Added the ability to record to a video programs run in runUtil with virtualX set. Also added videoUtil and imageUtil methods

+++ test/image.js
@@ -7,17 +7,24 @@ const tiptoe = require("tiptoe"),
 	imageUtil = require("../index").image;
 
 const FILES_DIR = path.join(__dirname, "files");
-
-console.log("This test will take a while.");
+const TEST_COMPRESS = false;
 
 tiptoe(
 	function getWH()
 	{
-		imageUtil.getWidthHeight(path.join(FILES_DIR, "input.png"), this);
+		imageUtil.getAutoCropDimensions(path.join(FILES_DIR, "test.png"), {cropColor : "#FFC0CB"}, this.parallel());
+		imageUtil.getAutoCropDimensions(path.join(FILES_DIR, "test2.png"), {cropColor : "#FFC0CB"}, this.parallel());
+		imageUtil.getWidthHeight(path.join(FILES_DIR, "input.png"), this.parallel());
 	},
-	function prepare(wh)
+	function prepare(cropInfo, cropInfo2, wh)
 	{
+		assert(Object.equals(cropInfo, {w : 1280, h : 720, x : 320, y : 180}));
+		assert(Object.equals(cropInfo2, {w : 640, h : 400, x : 0, y : 0}));
+
 		assert.strictEqual(JSON.stringify(wh), JSON.stringify([1487, 1500]));
+
+		if(!TEST_COMPRESS)
+			return this.finish();
 
 		fileUtil.unlink(path.join(FILES_DIR, "lossy.png"), this.parallel());
 		fileUtil.unlink(path.join(FILES_DIR, "lossless.png"), this.parallel());
@@ -44,7 +51,8 @@ tiptoe(
 			process.exit(1);
 		}
 
-		console.log("Now examine the files/*loss* files and visually compare to the input* files");
+		if(TEST_COMPRESS)
+			console.log("Now examine the files/*loss* files and visually compare to the input* files");
 
 		process.exit(0);
 	}
